% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-log-likelihoods.R
\name{accessTable}
\alias{accessTable}
\title{Function for accessing a double-KO contrast in a table}
\usage{
accessTable(table, double1, double2, single)
}
\arguments{
\item{table}{- the table to access}

\item{double1}{- an actor in the double KO}

\item{double2}{- another actor in the double KO}

\item{single}{- the singke KO/WT identifier}
}
\value{
the column in the table matching the description, NA if not found
}
\description{
Function for accessing a double-KO contrast in a table
}

