% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utils.R
\name{adjacencyToAncestry}
\alias{adjacencyToAncestry}
\title{Converts an adjacency matrix to an ancestry matrix}
\usage{
adjacencyToAncestry(adjacency)
}
\arguments{
\item{adjacency}{- the n x m adjacency matrix}
}
\value{
an n x m ancestry matrix
}
\description{
Converts an adjacency matrix (where element [i,j] is true when there is an edge from i to j) to an ancestry matrix
(where element [i,j] is true when i is an ancestor of j).
This function follows the convention that every node is considered its own ancestor.
}

