% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry-aggregation.R
\name{aggregateAncestries}
\alias{aggregateAncestries}
\title{Aggregate individual reporter-ancestries into one master ancestry}
\usage{
aggregateAncestries(ancestries)
}
\arguments{
\item{ancestries}{- a actors x actors+1 x reporters array of individual
reporter-ancestries.}
}
\value{
a single ancestry (with possibly duplicated actors) which is a matrix of
size (new actors) x (new actors + reporters)
}
\description{
Aggregate individual reporter-ancestries into one master ancestry
}

