% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{ancestries2graph}
\alias{ancestries2graph}
\title{A function to collapse the array of ancestries into one tree with
Possibly duplicate nodes.}
\usage{
ancestries2graph(ancestries, actors = dimnames(ancestries)[[1]],
  reporters = dimnames(ancestries)[[3]])
}
\description{
A function to collapse the array of ancestries into one tree with
Possibly duplicate nodes.
}

