% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-utils.R
\name{ancestryToAdjacency}
\alias{ancestryToAdjacency}
\title{Converts an ancestry matrix to an adjacency matrix}
\usage{
ancestryToAdjacency(ancestry)
}
\arguments{
\item{ancestry}{- the n x m ancestry matrix}
}
\value{
an n x m adjacency matrix
}
\description{
Converts an ancestry matrix (where element [i,j] is true when i is an ancestor of j)
to an adjacency matrix (where element [i,j] is true when there is an edge from i to j).
This function does not take self-ancestry seriously.
}

