% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-learner.R
\name{attachReporters}
\alias{attachReporters}
\title{Greedy reporter assignment}
\usage{
attachReporters(lll, actor.ancestry, debug.statements = FALSE)
}
\description{
Finds the maximum-scoring network with arcs to reporters for a fixed actor network
using ancestries.
}

