\name{bionetwork-package}
\alias{bionetwork-package}
\alias{bionetwork}
\docType{package}
\title{
\packageTitle{bionetwork}
}
\description{
\packageDescription{bionetwork}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bionetwork}
\packageIndices{bionetwork}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bionetwork}

Maintainer: \packageMaintainer{bionetwork}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
