% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{branchBoundAStar}
\alias{branchBoundAStar}
\title{Search for a reporter's best network}
\usage{
branchBoundAStar(lll, reporters = getReporters(lll),
  searchSelection = whichLastMax, debug.statements = FALSE)
}
\arguments{
\item{lll}{the LocalLogLikelihoods object}

\item{reporters}{the reporters for which we're building the network (all reporters in lll by default)}

\item{searchSelection}{function determining the search order, by default using lastMax,
which corresponds to DFS A*. An interesting alternative is choosing a random max,
choosing the first max will yield the abysmally slow BFS A*.}

\item{debug.statements}{whether to print out debug statements}
}
\value{
a list containing the ancestry structure specification (two adjacency matrices: one for present edges, and one for uncertain edges) and the reporter-wise score vector.
}
\description{
This version is a variation on A* search
}

