% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{branchBoundAStarOnTree}
\alias{branchBoundAStarOnTree}
\title{Search for a best network (without actor duplication)}
\usage{
branchBoundAStarOnTree(lll, reporters = getReporters(lll),
  searchSelection = whichLastMax, debug.level = 0)
}
\arguments{
\item{lll}{the LocalLogLikelihoods object}

\item{reporters}{the reporters for which we're building the network (all reporters in lll by default)}

\item{searchSelection}{function determining the search order, by default using lastMax,
which corresponds to DFS A*. An interesting alternative is choosing a random max,
choosing the first max will yield the abysmally slow BFS A*.}

\item{debug.level}{granularity of debug statements. 0 is none, larger numbers mean more statements.}
}
\value{
a list containing the ancestry structure specification (two adjacency matrices: one for present edges, and one for uncertain edges) and the reporter-wise score vector.
}
\description{
This version is a variation on A* search
}

