% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-with-splits.R
\name{cleanUpAncestry}
\alias{cleanUpAncestry}
\title{Clean up a single-reporter ancestry}
\usage{
cleanUpAncestry(ancestry)
}
\arguments{
\item{ancestry}{the ancestry matrix}
}
\value{
cleaned up ancestry matrix
}
\description{
Remove ancestry relationships among non-ancestors of the reporter
}

