% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{compute.score.change}
\alias{compute.score.change}
\title{Compute the score change}
\usage{
compute.score.change(reporter, old.ancestry, new.ancestry, old.actor.ancestry,
  new.actor.ancestry, laps, n = howManyActors(laps))
}
\description{
Compute the score change
}

