% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{edgesFromMatrix}
\alias{edgesFromMatrix}
\title{Utility for output
Convert adjacency matrix to cytoscape edge list}
\usage{
edgesFromMatrix(adjacency, nodeNames = if (!is.null(colnames(adjacency)))
  colnames(adjacency) else 1:max(dim(adjacency)))
}
\description{
Utility for output
Convert adjacency matrix to cytoscape edge list
}

