% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{getBestReporters}
\alias{getBestReporters}
\title{Greedy reporter assignment:
Finds the maximum-scoring network with arcs to reporters for a fixed actor network}
\usage{
getBestReporters(laps, nActors = howManyActors(laps),
  nReporters = howManyReporters(laps), network = matrix(FALSE, nrow =
  nActors, ncol = (nActors + nReporters)))
}
\description{
Greedy reporter assignment:
Finds the maximum-scoring network with arcs to reporters for a fixed actor network
}

