% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-helpers.R
\name{getChildNodesWide}
\alias{getChildNodesWide}
\title{Get children in the search graph (wide version)}
\usage{
getChildNodesWide(lll, search.node)
}
\arguments{
\item{lll}{the LocalLogLikelihoods object}

\item{search.node}{the search node whose children we're generating}
}
\value{
a list of search nodes (the children)
}
\description{
Get children in the search graph (wide version)
}

