% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-helpers.R
\name{getChildrenInToggleSearch}
\alias{getChildrenInToggleSearch}
\title{Toggle search children}
\usage{
getChildrenInToggleSearch(adjacency)
}
\value{
list of adjacency matrices resulting from toggling an edge
}
\description{
Powers an edge-toggling search over adjacency matrices
}

