% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-helpers.R
\name{getCost}
\alias{getCost}
\title{Extract the cost bounds from the node}
\usage{
getCost(search.node)
}
\arguments{
\item{search.node}{The search node object}
}
\value{
Cost bounds
}
\description{
We negate the scores because the search algorithm uses costs while we score based on likelihoods
}

