% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-with-splits.R
\name{getHeuristicScore}
\alias{getHeuristicScore}
\title{The heuristic + score for A*}
\usage{
getHeuristicScore(laps, reporterIndex, depth, adjacency)
}
\description{
The heuristic + score for A*
}

