% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-with-splits.R
\name{getHeuristicScoreOld}
\alias{getHeuristicScoreOld}
\title{The heuristic + score for A*}
\usage{
getHeuristicScoreOld(laps, reporterIndex, depth, adjacency)
}
\description{
The heuristic + score for A*
}

