% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-log-likelihoods.R
\name{getLogLikelihoods}
\alias{getLogLikelihoods}
\title{Function for getting log-likelihoods out of the fit}
\usage{
getLogLikelihoods(fit)
}
\arguments{
\item{fit}{- limma's fit object}
}
\value{
a list of two matrices representing the log likelihoods of equal and differential expression for each
reporter for each contrast in the fit.
}
\description{
Function for getting log-likelihoods out of the fit
}

