% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{getMLNetwork}
\alias{getMLNetwork}
\title{Greedy edge toggle search:}
\usage{
getMLNetwork(lp, nActors = length(getActors(lp)),
  nReporters = howManyReporters(lp), network = matrix(FALSE, nrow = nActors,
  ncol = (nActors + nReporters)), score = scoreNetwork(network, lp))
}
\description{
Greedy edge toggle search:
}

