% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{getNextGraphSearchStates}
\alias{getNextGraphSearchStates}
\title{Get the next state branches a graph search}
\usage{
getNextGraphSearchStates(adjacency, uncertain, fragile = TRUE)
}
\arguments{
\item{adjacency}{the present adjacency matrix}

\item{uncertain}{the present uncertain matrix}

\item{fragile}{whether to fail when there are no uncertain edges, (returns an empty
list if not fragile)}
}
\value{
a list of two elements (corresponding to making the last (or first)
ucertain edge certain) where each is the state description of the next branch,
consisting of an adjacency and an uncertain edge matrix
}
\description{
Get the next state branches a graph search
}

