% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{getNextSearchStates}
\alias{getNextSearchStates}
\title{Get the next state branches in the search}
\usage{
getNextSearchStates(adjacency, uncertain, last.edge = TRUE)
}
\arguments{
\item{adjacency}{the present adjacency matrix}

\item{uncertain}{the present uncertain matrix}

\item{last.edge}{whether to pick the last uncertain edge to make certain (default is
true), if false picks the first uncertain edge instead.}
}
\value{
a list of two elements (corresponding to making the last (or first)
ucertain edge certain) where each is the state description of the next branch,
consisting of an adjacency and an uncertain edge matrix
}
\description{
If the possible ancestry/nonancestry matrices are provided, the process is
sped up in some cases.
}

