% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-functions.R
\name{getScoreBounds}
\alias{getScoreBounds}
\title{Score bounds for underdetermined graph}
\usage{
getScoreBounds(lll, possible.ancestors, possible.nonancestors)
}
\arguments{
\item{lll}{- The LocalLogLikelihoods object}

\item{possible.ancestors}{- boolean matrix indicating possible ancestors}

\item{possible.nonancestors}{- boolean matrix indicating possible nonancestors}
}
\value{
A 2x|reporters| matrix, with the top row indicating upper and the bottom row indicating lower bounds on the score.
}
\description{
Computes the score bounds for an underdetermined graph described by two matrices,
possible.ancestors and possible.nonancestors. These describe ancestry relationships
derived from a graph with present, absent, and undetermined edges.
The number of reporters in the result matrix is based on the number of reporters in
the input ancestry matrices. It is assumed that the input matrices have the same
dimensions.
}

