% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{getUncertaintyMatrix}
\alias{getUncertaintyMatrix}
\title{Make uncertainty matrix from depth + dimension}
\usage{
getUncertaintyMatrix(depth = 0, n, m = n)
}
\arguments{
\item{depth}{- search depth}

\item{n}{- number of rows}

\item{m}{- number of columns}
}
\description{
Make uncertainty matrix from depth + dimension
}

