% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-functions.R
\name{groupReporters}
\alias{groupReporters}
\title{Get groups of reporters based on parent signatures.}
\usage{
groupReporters(network, name.groups = TRUE)
}
\description{
Get groups of reporters based on parent signatures.
}

