% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{inclusive.ancestry}
\alias{inclusive.ancestry}
\title{Inclusive ancestry of the square part of the matrix}
\usage{
inclusive.ancestry(network)
}
\description{
Inclusive ancestry of the square part of the matrix
}

