% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-helpers.R
\name{initSearchNode}
\alias{initSearchNode}
\title{Create the initial search node}
\usage{
initSearchNode(lll, reporters = getReporters(lll), use.sparse = TRUE)
}
\arguments{
\item{lll}{LocalLogLikelihoods object}

\item{reporters}{array of reporter names to use (optional, default pulled from lll)}
}
\value{
a search node (specified by adjacency and uncertainty matrices and a vector
of reporter contributions to the score)
}
\description{
Creates the initial (all-uncertain) node.
}

