% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-helpers.R
\name{isGoalNode}
\alias{isGoalNode}
\title{Goal node check}
\usage{
isGoalNode(search.node)
}
\arguments{
\item{search.node}{the search node}
}
\value{
TRUE iff this is a goal node
}
\description{
We reach a search goal when nothing is uncertain, or the upperbound is the lowerbound
}

