% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{log1mexp}
\alias{log1mexp}
\title{Compute log(1-exp(x)) accurately.
Based on "Accurately Computing log(1-exp(-|a|))" by Martin Mächler}
\usage{
log1mexp(x)
}
\description{
Compute log(1-exp(x)) accurately.
Based on "Accurately Computing log(1-exp(-|a|))" by Martin Mächler
}

