% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{logicVector}
\alias{logicVector}
\title{Make a logic vector}
\usage{
logicVector(n = 1, i = 1)
}
\arguments{
\item{n}{vector length}

\item{i}{the position of the true value}
}
\description{
Makes a vector of a cetain length with a true value in a certain position (remainder false)
}

