% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/using-limma.R
\name{makeLimmaLogProbs}
\alias{makeLimmaLogProbs}
\title{Use limma to get lprobs.}
\usage{
makeLimmaLogProbs(fit, actors, theColnames = colnames(fit$coefficients),
  doubleSpecs = {     ds = NULL     for (gene1 in actors) for (gene2 in
  actors) if ((str = paste0(gene1, gene2)) \%in\% theColnames)          ds =
  append(ds, list(list(str, c(gene1, gene2))))     ds }, wt = "WT",
  prior = 0.01)
}
\description{
Starting point is having a 'fit' (before contrasts).
Automatically build contrasts and run the right models.
}

