% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-log-likelihoods.R
\name{makeLLL}
\alias{makeLLL}
\title{Make a LocalLogLikelihoods object from a limma fit}
\usage{
makeLLL(fit, actors, wt = "WT")
}
\arguments{
\item{fit}{- the fit object produced by limma}

\item{actors}{- the list of actor names}

\item{wt}{- string specifying wild type}
}
\value{
a LocalLogLikelihoods object
}
\description{
Make a LocalLogLikelihoods object from a limma fit
}

