% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-helpers.R
\name{makeSearchNode}
\alias{makeSearchNode}
\title{Makes a search node}
\usage{
makeSearchNode(lll, adjacency, uncertain)
}
\arguments{
\item{lll}{LocalLogLikelihoods object}

\item{adjacency}{adjacency matrix}

\item{uncertain}{uncertainty matrix}
}
\value{
a search node (list made up of adjacency, uncertainty, and per-reporter
score vector)
}
\description{
Makes a search node
}

