% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{multiStartANetworkSearch}
\alias{multiStartANetworkSearch}
\title{Multiple start (by default greedy) search,
only over ancesty-like networks:}
\usage{
multiStartANetworkSearch(lp, searchFunction = getMLNetwork)
}
\description{
Multiple start (by default greedy) search,
only over ancesty-like networks:
}

