% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{multiStartNetworkSearch}
\alias{multiStartNetworkSearch}
\title{Multiple start (by default greedy) search:}
\usage{
multiStartNetworkSearch(lp, searchFunction = getMLNetwork)
}
\description{
Multiple start (by default greedy) search:
}

