% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-log-likelihoods.R
\docType{methods}
\name{+,LocalLogLikelihoods,LocalLogLikelihoods-method}
\alias{+,LocalLogLikelihoods,LocalLogLikelihoods-method}
\title{Overloaded + operator to merge LocalLogLikelihoods objects}
\usage{
\S4method{+}{LocalLogLikelihoods,LocalLogLikelihoods}(e1, e2)
}
\description{
Overloaded + operator to merge LocalLogLikelihoods objects
}

