% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{replaceNAs}
\alias{replaceNAs}
\title{Replaces all NAs in any indexable data structure (with 0 by default)}
\usage{
replaceNAs(x, replacement = 0)
}
\arguments{
\item{x}{- original}

\item{replacement}{- the thing with which to replace zeros}
}
\value{
version where all NAs were replaced by 0s
}
\description{
Replaces all NAs in any indexable data structure (with 0 by default)
}

