% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{score.edge.toggles.old}
\alias{score.edge.toggles.old}
\title{How to score an edge toggle}
\usage{
score.edge.toggles.old(network, lprobs, score = score.network(network,
  lprobs))
}
\description{
How to score an edge toggle
}

