% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{scoreEdgeToggles}
\alias{scoreEdgeToggles}
\title{Smart edge toggle
laps stands for local ancestry and pathway scores}
\usage{
scoreEdgeToggles(network, laps)
}
\description{
Smart edge toggle
laps stands for local ancestry and pathway scores
}

