% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-functions.R
\name{scoreLikelihoodsPerReporter}
\alias{scoreLikelihoodsPerReporter}
\title{Likelihood scores broken down by reporter}
\usage{
scoreLikelihoodsPerReporter(ancestry, lll)
}
\arguments{
\item{ancestry}{- the ancestry matrix}

\item{lll}{- the LocalLogLikelihoods object}
}
\value{
a vector with a log-likelihood for each reporter
}
\description{
Likelihood calculation for a full network, with the score contribution of each
reporter returned in a vector.
}

