% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{scoreNetwork}
\alias{scoreNetwork}
\title{Network score calculation
  Assumes the following structure for lprobs:
  lprobs$single.gt.wt is an n.actors x n.reporters matrix with the log probability
that the single KO has a greater effect than the WT effect on the reporter.
  lprobs$single.ngt.wt is log( 1 - exp( single.gt.wt ) )
  lprobs$double.eq.single is a 3-d n.actors (1) x n.actors (2) x n.reporters array
of the probability that actor 1 knocked out with actor 2 have the same effect on the
reporter as actor 1 alone.
  lprobs$double.gt.single is as above but probability of the double KO having a
greater effect}
\usage{
scoreNetwork(network, lp)
}
\description{
Network score calculation
  Assumes the following structure for lprobs:
  lprobs$single.gt.wt is an n.actors x n.reporters matrix with the log probability
that the single KO has a greater effect than the WT effect on the reporter.
  lprobs$single.ngt.wt is log( 1 - exp( single.gt.wt ) )
  lprobs$double.eq.single is a 3-d n.actors (1) x n.actors (2) x n.reporters array
of the probability that actor 1 knocked out with actor 2 have the same effect on the
reporter as actor 1 alone.
  lprobs$double.gt.single is as above but probability of the double KO having a
greater effect
}

