% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-helpers.R
\name{scoreSimply}
\alias{scoreSimply}
\title{Simple node scorer}
\usage{
scoreSimply(lll, adjacency)
}
\arguments{
\item{lll}{the LocalLogLikelihoods object}

\item{adjacency}{the adjacency matrix}
}
\description{
Simple node scorer
}

