% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R, R/search-with-splits.R
\name{searchWithSplits}
\alias{searchWithSplits}
\title{Exhaustive network search that will come up with actor-node-splits
if they are necessary based on the data.}
\usage{
searchWithSplits(laps, searchFunction = searchByAStar)

searchWithSplits(laps, searchFunction = searchByAStar)
}
\description{
Exhaustive network search that will come up with actor-node-splits
if they are necessary based on the data.

This version looks for the best actor ancestry individually per reporter and then
composes.
}

