% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-with-splits.R
\name{showUncertainAncestry}
\alias{showUncertainAncestry}
\title{A human-friendly representation of ancestry+uncertainty matrix}
\usage{
showUncertainAncestry(ancestry, uncertainty)
}
\arguments{
\item{ancestry}{the ancesty matrix}

\item{uncertainty}{the uncertainty matrix}
}
\value{
a char matrix with T/F in certain cells and ? in uncertain cells.
}
\description{
A human-friendly representation of ancestry+uncertainty matrix
}

