% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetwork.R
\name{signs.match}
\alias{signs.match}
\title{Data prep help: check that element-wise signs match on two vectors}
\usage{
signs.match(a, b)
}
\description{
Data prep help: check that element-wise signs match on two vectors
}

