% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-learner.R
\name{stripDots}
\alias{stripDots}
\title{Strip dots from strings}
\usage{
stripDots(strings)
}
\arguments{
\item{strings}{- the original string array}
}
\value{
strings with everything after the first dot (inclusive) stripped
}
\description{
Only keeps the part of each string that preceeds the first dot (.) character.
}

