% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-functions.R
\name{updateBounds}
\alias{updateBounds}
\title{Helper for updating upper/lower-bound matrix}
\usage{
updateBounds(bounds, scores)
}
\arguments{
\item{bounds}{- the bounds matrix}

\item{scores}{- vector of score candidates}
}
\value{
updated bounds where the "upper" is the max of scores+"upper" and "lower" is the min of scores+"lower"
}
\description{
Helper for updating upper/lower-bound matrix
}

