% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-with-splits.R
\name{whichLastMax}
\alias{whichLastMax}
\title{Given a vector, get the index of the last maximal element}
\usage{
whichLastMax(x)
}
\arguments{
\item{a}{vector of numbers}
}
\value{
the index of the last maximal element
}
\description{
Given a vector, get the index of the last maximal element
}

