% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-with-splits.R
\name{whichRandomLateMax}
\alias{whichRandomLateMax}
\title{Given a vector, get the index of a random maximal element}
\usage{
whichRandomLateMax(x, p = 0.9)
}
\arguments{
\item{a}{vector of numbers}
}
\value{
the index of a (non-uniformly) random maximal element
}
\description{
Given a vector, looks at its maximal elements, walking from last to first,
rejecting each with probability (1-p). (So has probability p to pick the last,
p*(1-p)th the 2nd last, etc. with the remainder of the probabilty mass placed
in the first).
}

